Return information about route matching state
Because react-router may want to perform redirections, or it may not
find the route, we need to deal with these scenarios. The JS functions
that are run will yield this information to us, and we simply pass these
upwards so that the Python WSGI framework can use the data to redirect
the user as required.